+++ src/core/crowdfunding.ts
@@ -1,4 +1,3 @@
-import { Proposal, Milestone } from '../models/types';
 import { Store } from '../models/Store';
 
 /**